1. adding BGQ and XE macros to test list 2. adding XE macro
Because I have not committed the change to the XT test to make it use the preprocessor, an XE will be detected as both an XT and XE.  This should not break anything for now because the XE config is identical to the XT config.

I failed to note previously that the importance of macro detection is that it can distinguish BGP and BGQ, which /bgsys probing cannot.