fix potential race issues with unique_tag
SAFE_MPI_GLOBAL_MUTEX is ***ONLY*** useful for serializing MPI calls
when MPI uses MPI_THREAD_SERIALIZED.  It preprocesses to ***NOTHING***
when MPI_THREAD_MULTIPLE is used.  That means that it will not protect
unique_tag's bounded increment all of the time.

To fix this, we allow the obolus(SafeMPI::charon) to be defined in
safempi.cc all of the time and use it explicitly in unique_tag.  When
used to protect MPI calls and ONLY to protect MPI calls, we continue to
use the SAFE_MPI_GLOBAL_MUTEX macro.

Signed-off-by: Jeff Hammond <jeff.science@gmail.com>