Add install instructions for OS X 10.8.
Add install instructions for OS X 10.8. The 10.6 instructions did not
work for me; `configure` indicated it did not detect a C++11 compiler,
which is nonsense, because Clang 3.4svn is C++11 compliant.

The problem appears to be in linking. If the "-stdlib=libc++" is not
specified, then Clang links with libstdc++. Since 10.8-vintage XCode
ships with a gcc 4.2-vintage version of libstdc++ that is not C++11
compliant, configure does not detect a C++11 compliant compiler.
Forcing Clang to link with libc++, which is C++11 compliant,
resolves the issue.