add C++ equivalent of INT_MAX
the safe downcast asserts check that the original type will not be
truncated on the downcast (if in fact long to libxsmm_blasint is an
int).

INT_MAX is the C way to do this.  i did it the C++11 way with auto and
std::numeric_limits to show that i am fancy.