First commit for adding in an NWChem interface for the MolDFT code.
Added in a new keyword "nwfile" to CalculationParameters.h to indicate the base file
name of the nwchem files. (Base file name as two files are needed: the output file
which contains the basis set information and the movecs file which contains the MO
coefficients. The software assumes an extension of .out for the former and an
extension of .movecs for the latter.) As of this commit, this keyword will print a
single line if provided in the input.

The newly added files are trimmed down versions of the Slymer package (not yet
publicly released) from mgr522 that contains all the moving parts to read an NWChem
output and movecs file and construct the orbitals from them.

Modified the CMakeLists.txt to include these new files. Autotools build files are
unchanged.